Add conftest for MyHelsinki Places importer
Refs AHTI-152
